Projects/YT/MVP1/requirements.txt
streamlit>=1.28.0
youtube-transcript-api>=0.6.2
google-generativeai>=0.3.1
python-dotenv>=1.0.0